Updated the "lgt2pdf.sh" and "lgt2txt.sh" POSIX shell scripts in order to prevent accidental deletion of the "logtal.dtd" and "logtalk.xsd" files in the Logtalk installation directory. Thanks to Michael Igler for the bug report.
git-svn-id: http://svn.logtalk.org/logtalk/trunk@5118 ac9c905b-388a-4ddf-9b6a-4063a022bc57

+++ xml/lgt2pdf.sh
@@ -164,7 +164,9 @@ else
 	echo
 fi
 
-rm -f logtalk.dtd
-rm -f logtalk.xsd
+if [ "$PWD" != "$LOGTALKHOME"/xml ] ; then
+	rm -f logtalk.dtd
+	rm -f logtalk.xsd
+fi
 
 exit 0
